New Generic Text class + Crafting Command
*Removed Guides, and Delve commands from LiftingBuildsModule.cs

*Created GenericCommandModule.cs and put Delve, Guides inside it

*Created !Crafting command that links to ghazzy's crafting video (Beta)
